clean_duplicate_descriptions/requirements.txt
google-auth==2.29.0
google-cloud-bigquery==3.25.0
protobuf==4.21.12